rename relocate srtEntries REgex
