+++ lib/parsers/srtEntries.js
@@ -8,7 +8,7 @@ const {
   emptyEntryRegex,
   validTimecodeRegex,
   invalidFirstEntryRegex,
-} = require('../shared/srtParseRegex');
+} = require('./srtEntriesRegex');
 
 function standardizeTimestamp(timestamp) {
   if (strictTimestampRegex.test(timestamp)) return timestamp;
