+++ lib/parsers/srtEntriesRegex.js
